Fixed piper - correct creation of process group and added semaphores for waiting the child processes.

+++ src/main/resources/piper.cpp
@@ -1,14 +1,17 @@
 #include<iostream>
+#include<unordered_set>
 #include<sys/wait.h>
+#include<sys/sem.h>
 #include<unistd.h>
 #include<string.h>
+#include<errno.h>
 #include<stdio.h>
 #include<string>
 #include<vector>
 using namespace std;
 void error_message (string s, int code) {
     cout << 0 << endl ;
-    cout << s << endl ;
+    cout << "Errno: " << errno << ", message: " << s << endl ;
     exit(code);
 }
 
@@ -18,7 +21,8 @@ struct pipes {
 };
 vector <int> fds;
 vector <int> children;
-int group=-1;
+int group=-1,sem_id;
+sembuf s;
 pair <int, int> make_process (char* grader) {
     pipes p;
     if ((pipe(p.pipeIn))||(pipe(p.pipeOut))) error_message("Pipe failed",-2);
@@ -32,17 +36,21 @@ pair <int, int> make_process (char* grader) {
         dup2(p.pipeOut[0],0);
         dup2(p.pipeIn[1],1);
         signal(SIGPIPE,SIG_IGN);
+        if (group==-1) setpgid(0,0);
+        else setpgid(0,group);
+        s.sem_op=1;
+        if (semop(sem_id,&s,1)==-1) error_message("Semop failed",-5);
         execl(grader,grader,NULL);
         error_message("Exec failed",-4);
     }
     else {
-        children.push_back(pid);
         if (group==-1) group=pid;
-        else setpgid(pid,group);
+        children.push_back(pid);
         close(p.pipeOut[0]); close(p.pipeIn[1]);
         return {p.pipeIn[0], p.pipeOut[1]};
     }
 }
+
 int count_digs (int num) {
     int cnt=0;
     for (;;) {
@@ -54,14 +62,18 @@ int count_digs (int num) {
 }
 int main (int argc, char* argv[]) {
     if (argc!=4) error_message("Arguments: number_of_grader_processes name_of_manager_program name_of_grader_program",-1);
+
     int processes=atoi(argv[1]);
-    signal(SIGPIPE,[] (int signal) {
-        error_message("Pipe error - violation of the protocol for communication!",0);
-    });
+    sem_id=semget(IPC_PRIVATE,1,0600|IPC_CREAT);
+    if (sem_id<0) error_message("Semget failed",-5);
+    s.sem_flg=0;
+    s.sem_num=0;
     for (int i=0; i<processes; i++) {
         char* grader;
         grader=argv[3];
         pair <int, int> res=make_process(grader);
+        s.sem_op=-1;
+        if (semop(sem_id,&s,1)==-1) error_message("Semop failed",-5);
         fds.push_back(res.first);
         fds.push_back(res.second);
     }
@@ -94,13 +106,13 @@ int main (int argc, char* argv[]) {
         if (WIFEXITED(status)) {
             if (WEXITSTATUS(status)) exit(WEXITSTATUS(status));
             else if (pid==manager) {
-                kill(-group,SIGTERM);
+                if (kill(-group,0)==0) kill(-group,SIGTERM);
                 exit(0);
             }
         }
         else if (WIFSIGNALED(status)) kill(getpid(),WTERMSIG(status));
         else if (WIFSTOPPED(status)) kill(getpid(),WSTOPSIG(status));
-        else error_message("Unknown status of child process - "+to_string(status),-5);
+        else error_message("Unknown status of child process - "+to_string(status),-6);
     }
     return 0;
 }
